[codex] notes foundation refactor (#81)
* refactor: decouple notes modules

* refactor: remove obsolete notes root module

* fix: address notes review feedback
